+++ src/_rules/focus-ring.js
@@ -1,6 +1,6 @@
 import { entriesToCss, escapeSelector } from '@unocss/core';
 
-// TODO: use actual variables and values when those has been defined
+// TODO: Remove deprecated fallback (--w-s-color-focused) in v2
 const focusRingStyle = entriesToCss(Object.entries({
   outline: '2px solid var(--w-s-color-border-focused, var(--w-s-color-focused))',
   'outline-offset': 'var(--w-outline-offset, 1px)',