+++ test/outline.js
@@ -5,14 +5,25 @@ import { lineWidth } from '#theme';
 setup();
 
 describe("outline", () => {
-  test("style", async ({ uno }) => {
-
+  test("supports setting outline styles", async ({ uno }) => {
     const classes = [
       "outline-none",
       "outline",
       "outline-dashed",
       "outline-dotted",
       "outline-double",
+    ];
+    const { css } = await uno.generate(classes);
+    expect(css).toMatchSnapshot();
+  });
+
+  test("supports setting outline colors", async ({ uno }) => {
+    const classes = [
+      "outline-inherit",
+      "outline-current",
+      "outline-transparent",
+      "outline-[--w-s-color-border]",
+      "outline-[var(--w-s-color-border)]",
     ];
     const { css } = await uno.generate(classes);
     expect(css).toMatchSnapshot();